+++ CHANGELOG.md
@@ -1,8 +1,12 @@
 # Sylver changelog
 
+- [[]()] Variable creation. Get and Set variables. 
+- [[3b11ae7](https://github.com/juliandavidmr/sylver/commit/3b11ae7fd9a70f2cea4b1e43f8c22ce29ced8664)] 
+Operate list numbers. `+`, `-`, `/`, `*` for array number.
+
 ## Version 1.0.2 #
-- [[]()] .
-- [[]()] CLI.
+- [[19c6757](https://github.com/juliandavidmr/sylver/commit/19c67570fde00bae8eecf564c779e86e60a573ee)] 
+CLI base.
 
 ## Version 1.0.1 #
 - [[de802a0](https://github.com/juliandavidmr/sylver/commit/b37538ebb08eaf251cc1b99038a2ca884be4c696)]